Clean code

+++ src/types/pet/pet.type.ts
@@ -24,7 +24,7 @@ export const PetSchema = z.object({
     name: z.string().min(1),
     species: z.string().min(1),
     breed: z.string().optional(),
-    age: z.coerce.number().int().nonnegative().optional(),
+    age: z.coerce.number().nonnegative().optional(),
     weight: z.coerce.number().nonnegative().optional(),
     imageUrl: z.string().optional(),
     ownerId: z.string().optional()
